Moved wrangling related code to separate module
